UHF-8242: Work around the map id count limit
