+++ src/js/react/apps/school-search/components/ResultsList.tsx
@@ -1,14 +1,11 @@
 import { SyntheticEvent, useState } from 'react';
 import { Button, IconMap, IconMenuHamburger, LoadingSpinner } from 'hds-react';
 import GlobalSettings from '../enum/GlobalSettings';
-import bucketToMap from '@/react/common/helpers/Aggregations';
-import { AggregationItem } from '@/types/Aggregation';
 import Result from '@/types/Result';
 import { School } from '../types/School';
 import ResultCard from './ResultCard';
 import Pagination from '@/react/common/Pagination';
-import CookieComplianceStatement from '@/react/common/CookieComplianceStatement';
-import ExternalLink from '@/react/common/ExternalLink';
+import ResultsMap from './ResultsMap';
 
 type ResultsListProps = {
   data: any;
@@ -18,30 +15,6 @@ type ResultsListProps = {
   page?: number;
   updatePage: Function
 }
-
-const getMapUrl = (ids?: AggregationItem[]) => {
-  const multipleBaseUrl = 'https://palvelukartta.hel.fi/fi/embed/search';
-  const singleBaseUrl = 'https://palvelukartta.hel.fi/fi/embed/unit';
-
-  if (!ids) {
-    return multipleBaseUrl;
-  }
-
-  const idMap = bucketToMap(ids);
-  const idArray = Array.from(idMap, (item) => item[0]);
-
-  if (idArray.length === 1) {
-    return `${singleBaseUrl}/${idArray[0]}`;
-  }
-
-  const mapUrl = new URL(multipleBaseUrl);
-  const params = new URLSearchParams();
-  params.set('units', idArray.join(','));
-
-  mapUrl.search = params.toString();
-
-  return mapUrl.toString();
-};
 
 const ResultsList = ({ data, error, isLoading, isValidating, page, updatePage }: ResultsListProps) => {
   const [useMap, setUseMap] = useState<boolean>(false);
@@ -65,38 +38,7 @@ const ResultsList = ({ data, error, isLoading, isValidating, page, updatePage }:
   const pages = Math.floor(total / size);
   const addLastPage = total > size && total % size;
 
-  const mapUrl = getMapUrl(data?.aggregations.ids.buckets);
   const showPagination = !useMap && (pages > 1 || addLastPage);
-
-  const getMap = () => {
-    if (Drupal.eu_cookie_compliance && Drupal.eu_cookie_compliance.hasAgreed('preference') && Drupal.eu_cookie_compliance.hasAgreed('statistics')) {
-      return (
-        <div className='school-search__map-container'>
-          <div className='unit-search__result--map'>
-            <iframe
-              title="Palvelukartta - Etusivu"
-              className="unit-search__map"
-              src={mapUrl}
-            >
-            </iframe>
-          </div>
-          <div className='unit-search__map-actions'>
-            <ExternalLink href={mapUrl} title={<span>{Drupal.t('Open large version of the map')}</span>} />
-          </div>
-        </div>
-      );
-    }
-
-    const url = new URL(mapUrl);
-
-    return (
-      <CookieComplianceStatement
-        host={url.host}
-        policyUrl={drupalSettings.helfi_school_search.cookie_privacy_url}
-        sourceUrl={mapUrl}
-      />
-    );
-  };
 
   return (
     <div className='react-search__results'>
@@ -126,7 +68,7 @@ const ResultsList = ({ data, error, isLoading, isValidating, page, updatePage }:
       </div>
       {
         useMap ?
-          getMap()
+          <ResultsMap ids={data?.aggregations?.ids?.buckets} />
         :
           <>
             {results.map((hit: Result<School>) => (
