+++ src/js/react/apps/school-search/components/ResultsMap.tsx
@@ -1,0 +1,77 @@
+import bucketToMap from '@/react/common/helpers/Aggregations';
+import CookieComplianceStatement from '@/react/common/CookieComplianceStatement';
+import ExternalLink from '@/react/common/ExternalLink';
+import { AggregationItem } from '@/types/Aggregation';
+
+type ResultsMapProps = {
+  ids?: AggregationItem[]
+}
+
+const ID_THRESHOLD = 90;
+
+const ResultsMap = ({ ids }: ResultsMapProps) => {
+  const multipleBaseUrl = 'https://palvelukartta.hel.fi/fi/embed/search';
+  const singleBaseUrl = 'https://palvelukartta.hel.fi/fi/embed/unit';
+
+  const getMapUrl = () => {
+    if (!ids) {
+      return multipleBaseUrl;
+    }
+  
+    const idMap = bucketToMap(ids);
+    const idArray = Array.from(idMap, (item) => item[0]);
+  
+    if (idArray.length === 1) {
+      return `${singleBaseUrl}/${idArray[0]}`;
+    }
+
+    const mapUrl = new URL(multipleBaseUrl);
+    const params = new URLSearchParams();
+
+    if (idArray.length > ID_THRESHOLD) {
+      params.set('service_node', '1100,1110,11187');
+      params.set('city', 'helsinki');
+      mapUrl.search = params.toString();
+
+      return mapUrl.toString();
+    }
+
+    params.set('units', idArray.join(','));
+  
+    mapUrl.search = params.toString();
+  
+    return mapUrl.toString();
+  };
+
+  const mapUrl = getMapUrl();
+
+  if (Drupal.eu_cookie_compliance && Drupal.eu_cookie_compliance.hasAgreed('preference') && Drupal.eu_cookie_compliance.hasAgreed('statistics')) {
+    return (
+      <div className='school-search__map-container'>
+        <div className='unit-search__result--map'>
+          <iframe
+            title="Palvelukartta - Etusivu"
+            className="unit-search__map"
+            src={mapUrl}
+          >
+          </iframe>
+        </div>
+        <div className='unit-search__map-actions'>
+          <ExternalLink href={mapUrl} title={<span>{Drupal.t('Open large version of the map')}</span>} />
+        </div>
+      </div>
+    );
+  }
+
+  const url = new URL(mapUrl);
+
+  return (
+    <CookieComplianceStatement
+      host={url.host}
+      policyUrl={drupalSettings.helfi_school_search.cookie_privacy_url}
+      sourceUrl={mapUrl}
+    />
+  );
+};
+
+export default ResultsMap;
